Fix address overflow in profile data collection.

+++ src/cpu/aarch64/vm/assembler_aarch64.hpp
@@ -544,7 +544,7 @@ class Address VALUE_OBJ_CLASS_SPEC {
 
   void lea(MacroAssembler *, Register) const;
 
-  static bool offset_ok_for_immed(int offset, int shift = 0) {
+  static bool offset_ok_for_immed(long offset, int shift = 0) {
     unsigned mask = (1 << shift) - 1;
     if (offset < 0 || offset & mask) {
       return (abs(offset) < (1 << (20 - 12))); // Unscaled offset
